rename module package to model package
