+++ accounting-backend/src/main/java/model/AccountEntry.java
@@ -1,4 +1,4 @@
-package module;
+package model;
 
 import java.time.LocalDate;
